add Dashboard related components in Dashboard folder
