+++ client/src/components/Dashboard/AddSchool.jsx
