+++ client/src/components/Dashboard/LoginForm.jsx
@@ -1,23 +1,23 @@
 import React from 'react';
-import logo from '../assets/SeriousFun_Logo.jpg';
+import logo from '../../assets/SeriousFun_Logo.jpg';
 
-const RegisterForm = props => {
+const LoginForm = props => {
   return (
     <div className='container'>
       <div className='row'>
         <div className='col-md-8 col-md-offset-2'>
           <img className='img-circle' width={350} src={logo} alt="logo"/> 
-          <h1>User Account Registration</h1><br/>
+          <h1>User Credentials</h1><br/>
           <div className='well'>
-            <form onSubmit={props.handleRegister}>
+            <form onSubmit={props.handleLogin}>
               <div className='form-group'>
                 <span className='input-group'>
                 <span className='input-group-addon main-color-bg'><i className="glyphicon glyphicon-user"></i></span>
                 <input type='text' 
                       className='form-control'
-                      placeholder='Please select a User Name'
-                      name='registerUsername'
-                      value={props.registerUsername}
+                      placeholder='Please enter your User Name'
+                      name='loginUsername'
+                      value={props.loginUsername}
                       onChange={props.handleInputChange}
                       required     
                 />
@@ -28,16 +28,19 @@ const RegisterForm = props => {
                 <span className='input-group-addon main-color-bg'><i className="glyphicon glyphicon-lock"></i></span>
                 <input type='password'
                       className='form-control'
-                      placeholder='Please select a password'
-                      name='registerPassword'
-                      value={props.registerPassword}
+                      placeholder='Please enter your password'
+                      name='loginPassword'
+                      value={props.loginPassword}
                       onChange={props.handleInputChange}
                       required
                 />
                 </span>
               </div>
+              <div className='checkbox'>
+                <label><input type="checkbox"/>Remember Me</label>
+              </div>
               <input type="submit"
-                    value='Register'
+                    value='Login'
                     className='btn btn-danger main-color-bg'
               />
             </form>
@@ -48,4 +51,4 @@ const RegisterForm = props => {
   )
 }
 
-export default RegisterForm;
+export default LoginForm;
