+++ client/src/components/Dashboard/RegisterForm.jsx
@@ -1,23 +1,23 @@
 import React from 'react';
-import logo from '../assets/SeriousFun_Logo.jpg';
+import logo from '../../assets/SeriousFun_Logo.jpg';
 
-const LoginForm = props => {
+const RegisterForm = props => {
   return (
     <div className='container'>
       <div className='row'>
         <div className='col-md-8 col-md-offset-2'>
           <img className='img-circle' width={350} src={logo} alt="logo"/> 
-          <h1>User Credentials</h1><br/>
+          <h1>User Account Registration</h1><br/>
           <div className='well'>
-            <form onSubmit={props.handleLogin}>
+            <form onSubmit={props.handleRegister}>
               <div className='form-group'>
                 <span className='input-group'>
                 <span className='input-group-addon main-color-bg'><i className="glyphicon glyphicon-user"></i></span>
                 <input type='text' 
                       className='form-control'
-                      placeholder='Please enter your User Name'
-                      name='loginUsername'
-                      value={props.loginUsername}
+                      placeholder='Please select a User Name'
+                      name='registerUsername'
+                      value={props.registerUsername}
                       onChange={props.handleInputChange}
                       required     
                 />
@@ -28,19 +28,16 @@ const LoginForm = props => {
                 <span className='input-group-addon main-color-bg'><i className="glyphicon glyphicon-lock"></i></span>
                 <input type='password'
                       className='form-control'
-                      placeholder='Please enter your password'
-                      name='loginPassword'
-                      value={props.loginPassword}
+                      placeholder='Please select a password'
+                      name='registerPassword'
+                      value={props.registerPassword}
                       onChange={props.handleInputChange}
                       required
                 />
                 </span>
               </div>
-              <div className='checkbox'>
-                <label><input type="checkbox"/>Remember Me</label>
-              </div>
               <input type="submit"
-                    value='Login'
+                    value='Register'
                     className='btn btn-danger main-color-bg'
               />
             </form>
@@ -51,4 +48,4 @@ const LoginForm = props => {
   )
 }
 
-export default LoginForm;
+export default RegisterForm;
